+++ client/src/input/Block.js
@@ -247,11 +247,17 @@ export default Garnish.Base.extend({
                     <li><a data-icon="field" data-action="copy" href="#" type="button" role="button" aria-label="${Craft.t('neo', 'Copy')}">${Craft.t('neo', 'Copy')}</a></li>
                     <li><a data-icon="brush" data-action="paste" href="#" type="button" role="button" aria-label="${Craft.t('neo', 'Paste')}">${Craft.t('neo', 'Paste')}</a></li>
                     <li><a data-icon="share" data-action="duplicate" href="#" type="button" role="button" aria-label="${Craft.t('neo', 'Clone')}">${Craft.t('neo', 'Clone')}</a></li>
-                  </ul>
+                  </ul>`)
+
+    if (type.isDeletableByUser()) {
+      elementHtml.push(`
                   <hr>
                   <ul class="padded">
                     <li><a class="error" data-icon="remove" data-action="delete" href="#" type="button" role="button" aria-label="${Craft.t('neo', 'Delete')}">${Craft.t('neo', 'Delete')}</a></li>
-                  </ul>
+                  </ul>`)
+    }
+
+    elementHtml.push(`
                 </div>
               </div>
             </div>
@@ -381,7 +387,14 @@ export default Garnish.Base.extend({
       }
 
       const detectChange = () => this._detectChange()
-      const observer = new window.MutationObserver(() => setTimeout(detectChange, 200))
+      const observer = new window.MutationObserver(() => {
+        setTimeout(detectChange, 200)
+
+        // Ensure blocks that are supposed to be non-editable by the user remain so
+        if (!this.getBlockType().isEditableByUser() && !this.$container.hasClass('is-disabled-for-user')) {
+          this.$container.addClass('is-disabled-for-user')
+        }
+      })
 
       observer.observe(this.$container[0], {
         attributes: true,