+++ client/src/input/Buttons.js
@@ -85,8 +85,8 @@ export default Garnish.Base.extend({
       const type = item.getType()
 
       if (type === 'blockType') {
-        // Ignore disabled block types
-        if (!item.getEnabled()) {
+        // Ignore disabled block types, or block types for which the current user isn't allowed to create blocks
+        if (!item.getEnabled() || !item.isCreatableByUser()) {
           continue
         }
 
@@ -145,8 +145,8 @@ export default Garnish.Base.extend({
       const type = item.getType()
 
       if (type === 'blockType') {
-        // Ignore disabled block types
-        if (!item.getEnabled()) {
+        // Ignore disabled block types, or block types for which the current user isn't allowed to create blocks
+        if (!item.getEnabled() || !item.isCreatableByUser()) {
           continue
         }
 
@@ -192,6 +192,19 @@ export default Garnish.Base.extend({
   initUi () {
     $('.menubtn', this.$container).menubtn()
     this.updateResponsiveness()
+
+    // If no buttons were rendered (e.g. if all valid block types are disabled for the user), hide the button container
+    if (this.$buttonsContainer.children().length === 0) {
+      const parent = this.$container.parent()
+      const grandParent = parent.parent()
+      const childrenContainer = grandParent.children('.ni_blocks')
+
+      if (childrenContainer.length === 0 || childrenContainer.children().length === 0) {
+        grandParent.addClass('hidden')
+      } else {
+        parent.addClass('hidden')
+      }
+    }
   },
 
   getBlockTypes () {