+++ src/migrations/Install.php
@@ -178,14 +178,10 @@ class Install extends Migration
     public function safeDown()
     {
         // Convert Neo fields to Matrix fields before dropping Neo tables
-        $fields = Craft::$app->getFields()->getAllFields();
-
-        foreach ($fields as $field) {
-            if ($field instanceof Field) {
-                // Don't bother deleting the old Neo block types and groups during the conversion, since we're about to
-                // drop the tables anyway
-                Neo::$plugin->conversion->convertFieldToMatrix($field, false);
-            }
+        foreach (Neo::$plugin->fields->getNeoFields() as $field) {
+            // Don't bother deleting the old Neo block types and groups during the conversion, since we're about to
+            // drop the tables anyway
+            Neo::$plugin->conversion->convertFieldToMatrix($field, false);
         }
 
         $this->dropTableIfExists('{{%neoblocks_owners}}');